Moved functions only used in one file there.
